// plugins/functions/mri-pa-config/src/types.ts
export interface IRequest extends Request {
    dbConnections: {
        db: ConnectionInterface;
    };
}
// ...
        interface Request {
            db: any;
        }
// ...
export interface IDBCredentialsType {
    database: string;
    schema?: string;
    dialect: string;
    host: string;
    port: number;
    user: string;
    password: string;
    max?: number | undefined;
    min?: number | undefined;
    idleTimeoutMillis?: number | undefined;
  }